postgres_to_es/sqlite_to_postgres/requirements.txt
django==3.2
python-dotenv==0.21.0
psycopg2-binary==2.9.3